+++ src/App.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
-import App from './presentation/App';
+import AppLayout from './presentation/AppLayout';
 
 test('renders learn react link', () => {
-  render(<App />);
+  render(<AppLayout />);
   const linkElement = screen.getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
 });
