+++ src/presentation/screens/categories/Categories.jsx
@@ -1,23 +1,26 @@
-import { Grid, Center, Stack, Button, Skeleton } from '@chakra-ui/react';
-import Category from './components/Category';
-import { useNavigate } from 'react-router-dom';
-import { useGetCategories } from 'infrastructure/services';
+import { Button, Center, Grid, Skeleton, Stack } from '@chakra-ui/react';
 import { useAppStore } from 'application/store';
+import { useGetCategories } from 'infrastructure/services';
 import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+import Category from './components/Category';
 
 function Categories() {
   const navigate = useNavigate();
-  const selectedCategory = useAppStore((store) => store.selectedCategory);
+
   const setCategory = useAppStore((store) => store.setCategory);
   const playSelectedCategory = useAppStore((store) => store.playSelectedCategory);
   const { data: categories, isLoading: isCategoriesLoading } = useGetCategories();
+
   const startPlaying = () => {
     playSelectedCategory();
     navigate('/questions');
   };
+
   useEffect(() => {
     setCategory(-1);
   }, []);
+
   return (
     <Center width="100%" h="100vh">
       <Stack direction="column" width="60%" gap={6}>
@@ -29,7 +32,7 @@ function Categories() {
             ))}
           </Grid>
         </Skeleton>
-        <Button isLoading={isCategoriesLoading} disabled={!selectedCategory} onClick={startPlaying}>
+        <Button isLoading={isCategoriesLoading} onClick={startPlaying}>
           Start
         </Button>
       </Stack>
