+++ src/presentation/screens/questions/components/QuestionTitle.jsx
@@ -1,6 +1,5 @@
-import PropTypes from 'prop-types';
-import React from 'react';
 import { Heading } from '@chakra-ui/react';
+import PropTypes from 'prop-types';
 
 function QuestionTitle({ title }) {
   return <Heading>{title}</Heading>;
