+++ src/presentation/screens/questions/components/Timer.jsx
@@ -1,7 +1,7 @@
-import PropTypes from 'prop-types';
-import React, { useEffect } from 'react';
 import { Heading } from '@chakra-ui/react';
 import { useAppStore } from 'application/store';
+import PropTypes from 'prop-types';
+import { useEffect } from 'react';
 
 function Timer({ skipQuestion }) {
   const timerCount = useAppStore((state) => state.timerCount);
